infrastructure/lims_rx_endpoint/project/lims_server/app_sm_rx/aaa_serializer_testbed/requirements.txt
flake8
Django
djangorestframework

